[opt] update preload to ts
